Separates caseFormats out into many files
