+++ src/fixtures/index.ts
@@ -1,1 +1,2 @@
 export * from './axiosResponse'
+export * from './stringCase'
