+++ src/utils/toSnakeCase.ts
@@ -1,0 +1,5 @@
+import { getWords } from './getWords'
+
+export function toSnakeCase(value: string): string {
+  return getWords(value).join('_')
+}
